requirements.txt
altgraph==0.17.4
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mysql-connector-python==9.2.0
numpy==2.2.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32-ctypes==0.2.3
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
